+++ szamlaagent/src/SzamlaAgent/SzamlaAgentException.php
@@ -2,6 +2,8 @@
 
 namespace SzamlaAgent;
 
+use Exception;
+
 /**
  * Számla Agent egyedi kivételeket kezelő osztály
  *
@@ -41,11 +43,11 @@ class SzamlaAgentException extends \Exception {
     /**
      * Számla Agent egyedi kivétel létrehozása
      *
-     * @param string     $message
-     * @param int        $code
-     * @param \Exception $previous
+     * @param string $message
+     * @param int $code
+     * @param Exception|null $previous
      */
-    public function __construct($message, $code = 0, \Exception $previous = null) {
+    public function __construct($message, $code = 0, Exception $previous = null) {
         parent::__construct($message, $code, $previous);
     }